+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+# v8.1.0
+
+- Added debug mode to help with troubleshooting and lower logging outputs.
+- Added Creative Mob Capture item to capture mobs in creative mode.
+- Added Enduring Capture Net to capture mobs with a higher durability.
+- Added Ironbound Containment Cage to capture mobs larger mobs.
+- Added Mystic Binding Crystal to capture very large and boss mobs.
+- Added wither specific drop like nether star and wither rose.
+- Improved blank mob capture card texture.
+- Improved translation and tooltip support.
+
 # v8.0.0
 
 First beta release of the next Easy Mob Farm 8.x for Fabric and Forge.